app/requirements.txt
fastapi
uvicorn[standard]
blinkpy
python-dotenv
httpx
aiofiles
